Fix transaction commit.
This is the hackiest of hacks, but it has to be done.  At all levels we
are checking types for the user so nothing should go wrong in theory.
This means that the type of entities below the TVar level is
essentially just an index over the log.  The logs and entries will
record the types they were instantiated at.  Because Swift is
applicative this means that we should have a stable identity and thus
the unsafe casts are always valid.